requirements.txt
numpy
pandas
matplotlib
datetime
pandas_datareader
keras
tensorflow
scikit-learn
streamlit
requests
pillow
streamlit-lottie